requirements.txt
# Core SQL and Database
sqlglot>=25.0.0
sqlparse>=0.5.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
PyMySQL>=1.1.1

# Data Generation and Manipulation
faker>=30.0.0
pandas>=2.2.3
numpy>=2.1.0

# AI and ML Components
sentence-transformers>=3.0.0
langchain>=0.3.0
langchain-community>=0.3.0
spacy>=3.8.0
anthropic>=0.37.0

# Vector Databases
faiss-cpu>=1.8.0
chromadb>=0.5.0

# UI Framework
gradio>=5.0.0
streamlit>=1.39.0

# Development and Testing
pytest>=8.0.0
black>=24.0.0

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.2
tqdm>=4.66.5
rich>=13.8.0
click>=8.1.7

# Export and Reporting
openpyxl>=3.1.5
matplotlib>=3.9.0
seaborn>=0.13.2
